+++ ValheimServersideQoL/ModConfig.CraftingStationsConfig.cs
@@ -6,6 +6,13 @@ partial record ModConfigBase
 {
     public sealed class CraftingStationsConfig(ConfigFile cfg, string section)
     {
+        public ConfigEntry<float> OpenClosestContainerRange { get; } = cfg.BindEx(section, 0f, """
+            The closest container within this range is automatically opened when a player interacts with a crafting station.
+            Opening a container this way refocuses the player's inventory as a side effect, so the player has to
+            manually switch to the crafting menu again to craft or repair items.
+            """);
+            
+
         public IReadOnlyDictionary<CraftingStation, StationCfg> StationConfig { get; } = new Func<IReadOnlyDictionary<CraftingStation, StationCfg>>(() =>
         {
             Dictionary<CraftingStation, bool> dict = new();
@@ -28,15 +35,15 @@ partial record ModConfigBase
         public sealed class StationCfg(ConfigFile cfg, string section, string prefix, CraftingStation station, bool hasExtensions)
         {
             public ConfigEntry<float>? BuildRange { get; } = station.m_areaMarker is null ? null :
-            cfg.Bind(section, $"{prefix}{nameof(BuildRange)}", station.m_rangeBuild, $"Build range of {(global::Localization.instance.Localize(station.m_name))}");
+                cfg.Bind(section, $"{prefix}{nameof(BuildRange)}", station.m_rangeBuild, $"Build range of {(global::Localization.instance.Localize(station.m_name))}");
             public ConfigEntry<float>? ExtraBuildRangePerLevel { get; } = station.m_areaMarker is null || !hasExtensions ? null :
-            cfg.Bind(section, $"{prefix}{nameof(ExtraBuildRangePerLevel)}", station.m_extraRangePerLevel, $"Additional build range per level of {(global::Localization.instance.Localize(station.m_name))}");
+                cfg.Bind(section, $"{prefix}{nameof(ExtraBuildRangePerLevel)}", station.m_extraRangePerLevel, $"Additional build range per level of {(global::Localization.instance.Localize(station.m_name))}");
             public ConfigEntry<float>? MaxExtensionDistance { get; } = !hasExtensions ? null :
-            cfg.Bind(section, $"{prefix}{nameof(MaxExtensionDistance)}", float.NaN, Invariant($"""
-                 Max distance an extension can have to the corresponding {(global::Localization.instance.Localize(station.m_name))} to increase its level.
-                 Increasing this range will only increase the range for already built extensions, you may need to temporarily place additional {(global::Localization.instance.Localize(station.m_name))} to be able to place the extension.
-                 {float.NaN} to use the game's default range. 
-                 """));
+                cfg.Bind(section, $"{prefix}{nameof(MaxExtensionDistance)}", float.NaN, Invariant($"""
+                     Max distance an extension can have to the corresponding {(global::Localization.instance.Localize(station.m_name))} to increase its level.
+                     Increasing this range will only increase the range for already built extensions, you may need to temporarily place additional {(global::Localization.instance.Localize(station.m_name))} to be able to place the extension.
+                     {float.NaN} to use the game's default range. 
+                     """));
         }
 
         static string NormalizeName(string name)